Add initial skeleton of a certmonger_getcert module
This is the really rough skeleton to begin building an ansible
module for using the certmonger 'getcert' utility to obtain X.509
certificates.  I have started on the option parsing only, and there
are still lots of TODO items there due to the large number of options
that 'getcert' has.  None of the functional part of the module has
been written yet.